chore(ECO-2872): Move `useLabelScrambler` to a sensible location
